Split out latex backend into a separate module
This commit splits out the latex circuit visualization module into a
separate module. This completes the decomposition of all the circuit
visualization backends into separate modules.

Fixes #1104
